Make code common between telemetry versions into common project
