--- dev/io.openliberty.microprofile.telemetry.1.0.internal/resources/META-INF/services/org.eclipse.microprofile.rest.client.spi.RestClientBuilderListener
@@ -1,1 +1,0 @@
-io.openliberty.microprofile.telemetry.internal.rest.TelemetryRestClientBuilderListener
