+++ dev/io.openliberty.microprofile.telemetry.common/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2023 IBM Corporation and others.
+# Copyright (c) 2022, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -10,30 +10,78 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
+openTelemetryVersion=1.19.0
+
 Bundle-Name: io.openliberty.microprofile.telemetry.common
 Bundle-SymbolicName: io.openliberty.microprofile.telemetry.common
-Bundle-Description: MicroProfile.telemetry.common, version 1.0
+Bundle-Description: MicroProfile.telemetry, version 1.0
 
 javac.source: 11
 javac.target: 11
 
 Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=11))"
 
-src: resources, src
+src: src, resources
+
+#-cdiannotations:
+
+-dsannotations: \
+  io.openliberty.microprofile.telemetry.common.internal.rest.TelemetryClientAsyncTaskWrapper,\
+  io.openliberty.microprofile.telemetry.common.internal.rest.TelemetryClientBuilderListener,\
+  io.openliberty.microprofile.telemetry.common.internal.rest.TelemetryServletContainerInitializer
+
+-dsannotations-inherit: true
 
 -includeresource: resources
 
+app-resources= \
+  META-INF/services/jakarta.ws.rs.ext.Providers | \
+  META-INF/services/org.eclipse.microprofile.rest.client.spi.RestClientBuilderListener
+
+Include-Resource: \
+  META-INF=resources/META-INF
+
+Service-Component: \
+    io.openliberty.microprofile.telemetry.internal.ResourceProvider; \
+    implementation:=com.ibm.wsspi.classloading.ResourceProvider; \
+    provide:=com.ibm.wsspi.classloading.ResourceProvider; \
+    configuration-policy:=optional; \
+	properties:="resources=${app-resources}" 
+
+# Need an explicit import for api.common because it gets used by weld proxies
+# created for other API classes
 Import-Package: \
+  io.opentelemetry.api.common,\
+  io.openliberty.microprofile.telemetry.common.*,\
   com.ibm.websphere.ras,\
   com.ibm.websphere.ras.annotation,\
   *
 
 Export-Package: \
-  io.openliberty.microprofile.telemetry.common.resources;version=1.0,\
-  io.openliberty.microprofile.telemetry.common.rest;version=1.0
+	io.openliberty.microprofile.telemetry.common.rest;version='1.0',\
+	io.openliberty.microprofile.telemetry.common.resources;version='1.0',\
+	io.openliberty.microprofile.telemetry.common.internal.cdi;version='1.0',\
+	io.openliberty.microprofile.telemetry.common.internal.helper;version='1.0',\
+	io.openliberty.microprofile.telemetry.common.internal.rest;version='1.0';thread-context=true
 
-WS-TraceGroup: TELEMETRY
+Private-Package: \
+	io.openliberty.microprofile.telemetry.common.internal.helper
 
 -buildpath: \
-    com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-    com.ibm.wsspi.org.osgi.service.component.annotations;version=latest
+    io.openliberty.jakarta.restfulWS.3.1;version=latest,\
+    io.openliberty.jakarta.annotation.2.1;version=latest,\
+    io.openliberty.jakarta.interceptor.2.1;version=latest,\
+    io.openliberty.jakarta.cdi.4.0;version=latest,\
+    io.openliberty.jakarta.servlet.6.0;version=latest,\
+    com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
+    com.ibm.ws.logging.core;version=latest,\
+    com.ibm.ws.cdi.interfaces.jakarta;version=latest,\
+    com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
+    io.openliberty.io.opentelemetry;version=latest,\
+    io.openliberty.org.eclipse.microprofile.config.3.0,\
+    io.openliberty.org.eclipse.microprofile.rest.client.3.0,\
+    io.openliberty.org.jboss.resteasy.common.ee10;version=latest,\
+    com.ibm.websphere.org.osgi.core;version=latest,\
+    com.ibm.ws.container.service;version=latest
+
+WS-TraceGroup: TELEMETRY
