+++ dev/io.openliberty.microprofile.telemetry.common/resources/META-INF/services/org.eclipse.microprofile.rest.client.spi.RestClientBuilderListener
@@ -1,0 +1,1 @@
+io.openliberty.microprofile.telemetry.common.internal.rest.TelemetryRestClientBuilderListener
